修改Y轴数值BUG

+++ JKChartView/JKChartView/JKChartView.m
@@ -263,10 +263,10 @@ static const CGFloat ProgressColorInfoViewWidth = 110.0;
 
 - (void)setYMaxValue:(double)maxValue {
     _yMaxValue = maxValue;
-    NSInteger yDivideValue = maxValue/(self.yLabelViews.count + 1);
+    double yDivideValue = maxValue/(self.yLabelViews.count + 1);
     for (int i = 0; i < self.yLabelViews.count; i++) {
         UILabel *yLabel = self.yLabelViews[self.yLabelViews.count - 1 - i];
-        yLabel.text = [NSString stringWithFormat:@"%ld", yDivideValue*(i+1)];
+        yLabel.text = [NSString stringWithFormat:@"%.1f", yDivideValue*(i+1)];
     }
 }
